refactor(GODT-1757): Strong Type IMAP Sequence
This patch also ensures that these this type is represented by uint32
which matches the spec's requirements.

+++ internal/state/responders.go
@@ -123,7 +123,7 @@ func (u *targetedExists) handle(snap *snapshot, stateID StateID) ([]response.Res
 			}
 		}
 
-		res = append(res, response.Recent().WithCount(recent))
+		res = append(res, response.Recent().WithCount(uint32(recent)))
 	}
 
 	return res, dbUpdate, nil
